Fixed a bug in Maya.fix_reference_namespace() method
The bug was about trying to query edits on a node which has a name with
unwanted characters in which Maya was raising UnicodeEncodeError. So, with
its final version Maya.fix_reference_namespace() skips this kind of
nodes, so it will not be able to keep the changes while updating the
namespace, which is ok, cause it is super rare to have a node name with
some unwanted characters.