Publishers are now working.
Publishers are a way to check the quality of the published files. To check the quality of a certain file you can register a function or a method as a publisher for a certain task type. Then it will be called by the environment before saving the file as a new version.